@AdamHillier 2020-09-04
Wow that did not work how I expected it would. This was a PR into the `larq:optimised-int8-kernel` branch (the #492 PR branch), and as a result of that PR being merged the branch got deleted and so this PR got automatically closed. I thought usually GitHub was clever enough to 'redirect' the PR to point to the master branch, but apparently not if it's a PR from a fork.